+++ apps/web/app/lol/champions/page.tsx
@@ -103,7 +103,7 @@ export default async function ChampionsPage({
     <div className="grid gap-8">
       <header className="page-hero p-5 md:p-8">
         <p className="type-kicker text-muted">League Directory</p>
-        <h1 className="type-title mt-3 sm:text-[2.4rem]">
+        <h1 className="type-page-title mt-3">
           Champions
         </h1>
         <p className="type-ui mt-3 text-fg/75">
