+++ apps/web/app/lol/pro-builds/page.tsx
@@ -192,7 +192,7 @@ export default async function ProBuildsIndexPage({
     <div className="grid gap-8">
       <header className="page-hero p-5 md:p-8">
         <p className="type-kicker text-muted">Tracked Matches</p>
-        <h1 className="type-title mt-3 sm:text-[2.4rem]">
+        <h1 className="type-page-title mt-3">
           Pro Builds
         </h1>
         <p className="type-ui mt-3 text-fg/75">
