+++ apps/web/app/tft/champions/page.tsx
@@ -13,7 +13,7 @@ export default async function TftChampionsPage() {
     <div className="grid gap-6">
       <section className="page-hero p-6">
         <p className="type-kicker text-muted">TFT Catalog</p>
-        <h1 className="type-title mt-3 sm:text-[2.4rem]">TFT Units</h1>
+        <h1 className="type-page-title mt-3">TFT Units</h1>
         <p className="type-ui mt-3 text-fg/75">Browse every unit in the live set and jump into quick detail pages.</p>
       </section>
       <TftCatalogGrid items={champions} basePath="/tft/champions" />
