+++ apps/web/app/tft/items/[itemId]/page.tsx
@@ -43,7 +43,7 @@ export default async function TftItemDetailPage({
         )}
         <div>
           <p className="type-kicker text-muted">TFT Item</p>
-          <h1 className="type-title mt-2 sm:text-[2.2rem]">{entity.name}</h1>
+          <h1 className="type-page-title mt-2">{entity.name}</h1>
           <p className="type-ui mt-2 text-muted">Item details</p>
         </div>
       </div>
